добавляет MovieController, реализует логику рендера карточки фильмов через него
